Move String extensions to separate file
Makes it more logical if you access these methods from other places in
the app
